Refactor queue management with modular service structure
- Split queue-related code into separate files under `services/queue/`
- Create `common.ts` with shared queue utilities and configuration
- Separate track and WAV conversion queues into dedicated files
- Improve code organization and maintainability
- Reduce code duplication across queue processing logic
